Add a port of enso_sst_covariance.ipynb to Chapel
This is an initial port of the simple ocean modeling Python code
enso_sst_covariance.ipynb to Chapel. It is a pretty direct translation and
has not been optimized in any significant way.

Because there isn't currently a Chapel interface to matplotlib.pyplot, it just
prints the computed array to stdout instead of displaying the final image.

Compile with:
chpl --fast enso_sst_covariance.chpl